+++ client/src/componentes/Paginate/Paginate.module.css
@@ -1,58 +1,47 @@
-.cardContainer {
-    border: none;
-    background-color: beige;
-    width: 220px;
-    height: 300px;
-    display: flex;
-    flex-direction: column;
-}
-    /* justify-content: center; /
-    align-items: center;
-}
-
-.imgCard {
-    transition: transform .2s; 
-}
-
-.imgCard:hover {
-    transform: scale(1.1);
-}
-
-.cardContainer:hover {
-    / transform: translateY(-0.4rem); /
-    / box-shadow: 3px 3px 5px 0px rgba(0,0,0,0.25); /
-    transition: all 0.4s;
-    border: thin solid black;
-    cursor: pointer;
-    box-sizing: border-box;
-}
-
-/ .cardInfo {
-    margin: 5px;
-} */
-
-.imgContainer {
-    background-color: #fff;
+.paginadoContainer {
+    margin-bottom: 15px;
+    margin-top: 15px;
     width: 100%;
-    height: 350px;
     display: flex;
     justify-content: center;
-    align-items: center;
 }
 
-.pokeType {
-    background-color: #3c5aa6cc;
-    margin: 0;
-    width: 100%;
-    height: 35px;
-    color: #ffcb05bb;
-    font-size: 1rem;
-    font-weight: 500;
-    line-height: 30px;
-}
-
-.pokeTypeContainer {
-    margin: 0;
+.paginadoBotonContainer {
+    width: 135px;
     display: flex;
-    width: 100%;
+    justify-content: space-evenly;
+}
+
+.paginadoBoton {
+    background-color: #fff;
+    border: 1.5px solid #000;
+    color: #000;
+    font-weight: bold;
+}
+
+/* .paginadoBoton:visited {
+    transform: translateY(-0.25rem);
+    transition: all 0.4s;
+    cursor: pointer;
+    color: #fff;
+    background-color: rgba(128, 128, 128, 0.89);
+    box-shadow: 1px 2px 1px 0 rgb(60, 60, 60);
+} */
+
+.paginadoBoton:hover {
+    transform: translateY(-0.25rem);
+    transition: all 0.4s;
+    cursor: pointer;
+    color: #fff;
+    background-color: rgba(128, 128, 128, 0.89);
+    box-shadow: 1px 2px 1px 0 rgb(60, 60, 60);
+}
+
+.paginadoSelect {
+    transform: translateY(-0.25rem);
+    transition: all 0.4s;
+    cursor: pointer;
+    color: #fff;
+    background-color: rgba(128, 128, 128, 0.89);
+    box-shadow: 1px 2px 1px 0 rgb(60, 60, 60);
 }
